+++ src/app/page-about/page-licenses/page-licenses-table/page-licenses-table.component.ts
