+++ src/app/page-qr-scanner-public/page-qr-scanner-routing.module.ts
