+++ src/app/page-qr-scanner-public/page-qr-scanner.module.ts
