+++ src/app/page-qr-scanner-public/page-qr-scanner.page.html
