+++ src/app/page-restricted-area/page-attendance-collect/components/event-list/event-list.component.ts
