+++ src/app/page-restricted-area/page-attendance-collect/page-scanner/scanner.page.ts
