+++ src/app/page-restricted-area/page-restricted-area-routing.module.ts
@@ -11,14 +11,14 @@ const routes: Routes = [
   {
     path: 'coletar-presenca',
     loadChildren: () =>
-      import('../page-attendance-collect/page-attendance-collect.module').then(
+      import('./page-attendance-collect/page-attendance-collect.module').then(
         (m) => m.PageAttendanceCollectPageModule
       ),
   },
   {
     path: 'gerenciar-admins',
     loadChildren: () =>
-      import('../page-manage-admins/page-manage-admins.module').then(
+      import('./page-manage-admins/page-manage-admins.module').then(
         (m) => m.PageManageAdminsPageModule
       ),
   },
