+++ src/app/tab-calendar/tab-calendar-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { PageCalendarEventPage } from '../page-calendar-event/page-calendar-event.page';
+import { PageCalendarEventPage } from './page-calendar-event/page-calendar-event.page';
 import { TabCalendarPage } from './tab-calendar.page';
 
 const routes: Routes = [
@@ -18,7 +18,7 @@ const routes: Routes = [
   {
     path: 'evento/:id',
     loadChildren: () =>
-      import('../page-calendar-event/page-calendar-event.module').then(
+      import('./page-calendar-event/page-calendar-event.module').then(
         (m) => m.PageCalendarEventPageModule
       ),
   },
